venv_requirements.txt
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.2
parso==0.5.1
pexpect==4.7.0
Pillow==6.1.0
pkg-resources==0.0.0
ptyprocess==0.6.0
pyparsing==2.4.2
python-dateutil==2.8.0
pytorch-utils==0.1
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.8
torch==1.2.0
torchvision==0.4.0
tqdm==4.36.1
wcwidth==0.1.7
Werkzeug==0.15.6
WTForms==2.2.1
